# backend-service/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Manajemen Anime Watchlist API",
        "description": "Dokumentasi lengkap REST API untuk Proyek Akhir Teknologi Integrasi Sistem. API ini menyediakan fitur autentikasi JWT, manajemen watchlist anime, pencarian anime via AniDB, dan manajemen user oleh admin.",
        "contact": {
            "name": "Raka Paksi",
            "email": "rakapaksi@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001/api",
            "description": "Backend Service Server"
        }
    ],
    "paths": {
...
        "/watchlists": {
            "get": {
                "tags": [
                    "Watchlist"
                ],
                "summary": "Get All Watchlists",
                "description": "Mengambil semua data watchlist milik user yang login. Admin dapat melihat seluruh watchlist dari semua user.",
                "operationId": "getWatchlists",
                "responses": {
                    "200": {
                        "description": "Daftar watchlist berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar watchlist berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "anime_id": {
                                                        "type": "string",
                                                        "example": "20"
                                                    },
                                                    "anime_title": {
                                                        "type": "string",
                                                        "example": "Naruto"
                                                    },
                                                    "anime_rating": {
                                                        "type": "string",
                                                        "example": "8.25"
                                                    },
                                                    "anime_image": {
                                                        "type": "string",
                                                        "example": "https://cdn.myanimelist.net/images/anime/13/17405.jpg"
                                                    },
                                                    "status_watching": {
                                                        "type": "string",
                                                        "example": "Watching"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid atau expired"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Watchlist"
                ],
                "summary": "Add to Watchlist",
                "description": "Menambahkan anime ke watchlist user yang sedang login.",
                "operationId": "addToWatchlist",
                "requestBody": {
                    "description": "Data anime yang akan ditambahkan ke watchlist",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "anime_id",
                                    "anime_title"
                                ],
                                "properties": {
                                    "anime_id": {
                                        "description": "ID anime dari AniDB",
                                        "type": "string",
                                        "example": "20"
                                    },
                                    "anime_title": {
                                        "description": "Judul anime",
                                        "type": "string",
                                        "example": "Naruto"
                                    },
                                    "anime_rating": {
                                        "description": "Rating anime",
                                        "type": "string",
                                        "example": "8.25"
                                    },
                                    "anime_image": {
                                        "description": "URL gambar anime",
                                        "type": "string",
                                        "example": "https://cdn.myanimelist.net/images/anime/13/17405.jpg"
                                    },
                                    "status_watching": {
                                        "description": "Status: Plan to Watch, Watching, Completed",
                                        "type": "string",
                                        "example": "Plan to Watch"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Anime berhasil ditambahkan ke watchlist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Anime berhasil ditambahkan ke watchlist"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "anime_id": {
                                                    "type": "string",
                                                    "example": "20"
                                                },
                                                "anime_title": {
                                                    "type": "string",
                                                    "example": "Naruto"
                                                },
                                                "anime_rating": {
                                                    "type": "string",
                                                    "example": "8.25"
                                                },
                                                "status_watching": {
                                                    "type": "string",
                                                    "example": "Plan to Watch"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The anime_id field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid atau expired"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT yang didapat dari endpoint /login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}